Last stable

+++ src/modules/user/routers.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException
 
 from src.core.dependencies import get_jwt_service, get_authentication_header, get_user_service
 from src.modules.jwt.schemas import JWTResponseDTO
-from src.modules.jwt.service import JWTService
+from src.modules.jwt.services import JWTService
 from src.modules.jwt.util import TokenType
 from src.modules.user.schemas import UserRegisterDTO, UserResponseDTO, UserLoginDTO
 from src.modules.user.services import UserService
@@ -18,6 +18,10 @@ async def get_user(
         jwt_service: JWTService = Depends(get_jwt_service)
 ) -> UserResponseDTO:
     payload = jwt_service.decode_token(token=user_jwt)
+    if payload.get("type"):
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    if payload.get("type") != "access":
+        raise HTTPException(status_code=401, detail="Unauthorized")
     if not payload:
         raise HTTPException(status_code=401, detail="Unauthorized")
 
@@ -81,7 +85,33 @@ async def delete_user(
     payload = jwt_service.decode_token(token=user_jwt)
     if not payload:
         raise HTTPException(status_code=401, detail="Unauthorized")
+    if payload.get("type"):
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    if payload.get("type") != "access":
+        raise HTTPException(status_code=401, detail="Unauthorized")
     if int(payload.get("sub", -1)) != user_id:
         raise HTTPException(status_code=403, detail="Not enough rights")
     user_service.delete_user(user_id)
     return None
+
+
+@user_router.get("/refresh")
+async def refresh(
+        user_jwt: str = Depends(get_authentication_header),
+        jwt_service: JWTService = Depends(get_jwt_service),
+        user_service: UserService = Depends(get_user_service)
+) -> JWTResponseDTO:
+    payload = jwt_service.decode_token(user_jwt)
+    if not payload:
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    if not payload.get("type"):
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    if payload.get("type") != "refresh":
+        raise HTTPException(status_code=401, detail="Unauthorized")
+
+    user = user_service.get_user_by_id(int(payload.get("sub"), -1))
+    return JWTResponseDTO(
+        user_id=user.id,
+        access=jwt_service.create_token(user.id, user.position, TokenType.ACCESS),
+        refresh=jwt_service.create_token(user.id, user.position, TokenType.REFRESH)
+    )
